+++ src/algorithms/libs/rtklib/rtklib_pntpos.cc
@@ -22,17 +22,23 @@
  * -----------------------------------------------------------------------------
  * Copyright (C) 2007-2013, T. Takasu
  * Copyright (C) 2017, Javier Arribas
- * Copyright (C) 2017, Carles Fernandez
+ * Copyright (C) 2017-2023, Carles Fernandez
  * All rights reserved.
  *
  * SPDX-License-Identifier: BSD-2-Clause
  * -----------------------------------------------------------------------------
  */
 
+#if ARMA_NO_BOUND_CHECKING
+#define ARMA_NO_DEBUG 1
+#endif
+
 #include "rtklib_pntpos.h"
 #include "rtklib_ephemeris.h"
 #include "rtklib_ionex.h"
 #include "rtklib_sbas.h"
+#include <armadillo>
+#include <cmath>
 #include <cstring>
 #include <vector>
 
@@ -599,6 +605,81 @@ int valsol(const double *azel, const int *vsat, int n,
 }
 
 
+// Lorentz inner product
+double lorentz(const arma::vec &x, const arma::vec &y)
+{
+    double p = x(0) * y(0) + x(1) * y(1) + x(2) * y(2) - x(3) * y(3);
+    return p;
+}
+
+
+// Bancroft method (see https://gssc.esa.int/navipedia/index.php/Bancroft_Method)
+// without travel time rotation
+arma::vec rough_bancroft(const arma::mat &B_pass)
+{
+    const int m = B_pass.n_rows;
+    arma::vec pos = arma::zeros<arma::vec>(4);
+    arma::mat BBB;
+    bool success;
+    if (m > 4)
+        {
+            success = arma::pinv(BBB, B_pass);
+        }
+    else
+        {
+            success = arma::inv(BBB, B_pass);
+        }
+    if (!success)
+        {
+            return pos;
+        }
+    const arma::vec e = arma::ones<arma::vec>(m);
+    arma::vec alpha = arma::zeros<arma::vec>(m);
+    for (int i = 0; i < m; i++)
+        {
+            arma::vec Bi = B_pass.row(i).t();
+            alpha(i) = lorentz(Bi, Bi) / 2.0;
+        }
+    const arma::vec BBBe = BBB * e;
+    const arma::vec BBBalpha = BBB * alpha;
+    const double a = lorentz(BBBe, BBBe);
+    const double b = lorentz(BBBe, BBBalpha) - 1.0;
+    const double c = lorentz(BBBalpha, BBBalpha);
+    const double root = std::sqrt(b * b - a * c);
+    arma::vec r(2);
+    r(0) = (-b - root) / a;
+    r(1) = (-b + root) / a;
+    arma::mat possible_pos = arma::zeros<arma::mat>(4, 2);
+    for (int i = 0; i < 2; i++)
+        {
+            possible_pos.col(i) = r(i) * BBBe + BBBalpha;
+            possible_pos(3, i) = -possible_pos(3, i);
+        }
+    arma::vec abs_omc(2);
+    for (int j = 0; j < m; j++)
+        {
+            for (int i = 0; i < 2; i++)
+                {
+                    const double c_dt = possible_pos(3, i);
+                    const double calc = arma::norm(B_pass.row(j).head(3).t() - possible_pos.head_rows(3).col(i)) + c_dt;
+                    const double omc = B_pass(j, 3) - calc;
+                    abs_omc(i) = std::abs(omc);
+                }
+        }
+
+    if (abs_omc(0) > abs_omc(1))
+        {
+            pos = possible_pos.col(1);
+        }
+    else
+        {
+            pos = possible_pos.col(0);
+        }
+
+    return pos;
+}
+
+
 /* estimate receiver position ------------------------------------------------*/
 int estpos(const obsd_t *obs, int n, const double *rs, const double *dts,
     const double *vare, const int *svh, const nav_t *nav,
@@ -630,6 +711,34 @@ int estpos(const obsd_t *obs, int n, const double *rs, const double *dts,
     for (i = 0; i < 3; i++)
         {
             x[i] = sol->rr[i];
+        }
+
+    // Rough first estimation to initialize the algorithm
+    if (opt->bancroft_init && (std::sqrt(x[0] * x[0] + x[1] * x[1] + x[2] * x[2]) < 0.1))
+        {
+            arma::mat B = arma::mat(n, 4, arma::fill::zeros);
+            for (i = 0; i < n; i++)
+                {
+                    B(i, 0) = rs[0 + i * 6];
+                    B(i, 1) = rs[1 + i * 6];
+                    B(i, 2) = rs[2 + i * 6];
+                    if (obs[i].code[0] != CODE_NONE)
+                        {
+                            B(i, 3) = obs[i].P[0];
+                        }
+                    else if (obs[i].code[1] != CODE_NONE)
+                        {
+                            B(i, 3) = obs[i].P[1];
+                        }
+                    else
+                        {
+                            B(i, 3) = obs[i].P[2];
+                        }
+                }
+            arma::vec pos = rough_bancroft(B);
+            x[0] = pos(0);
+            x[1] = pos(1);
+            x[2] = pos(2);
         }
 
     for (i = 0; i < MAXITR; i++)