+++ src/tests/unit-tests/signal-processing-blocks/filter/pulse_blanking_filter_test.cc
@@ -43,26 +43,28 @@ DEFINE_int32(pb_filter_test_nsamples, 1000000, "Number of samples to filter in t
 class PulseBlankingFilterTest : public ::testing::Test
 {
 protected:
-    PulseBlankingFilterTest() : item_size(sizeof(gr_complex)), nsamples(FLAGS_pb_filter_test_nsamples)
+    PulseBlankingFilterTest() : item_size(sizeof(gr_complex)),
+                                nsamples(FLAGS_pb_filter_test_nsamples)
     {
         queue = std::make_shared<Concurrent_Queue<pmt::pmt_t>>();
         config = std::make_shared<InMemoryConfiguration>();
     }
-    ~PulseBlankingFilterTest() override = default;
-    bool stop = false;
-    std::thread ch_thread;
+
     void start_queue();
     void wait_message();
     void process_message();
     void stop_queue();
     void init();
     void configure_gr_complex_gr_complex();
+
+    std::thread ch_thread;
     std::shared_ptr<Concurrent_Queue<pmt::pmt_t>> queue;
-    gr::top_block_sptr top_block;
     std::shared_ptr<InMemoryConfiguration> config;
+    gr::top_block_sptr top_block;
+    pmt::pmt_t message;
     size_t item_size;
     int nsamples;
-    pmt::pmt_t message;
+    bool stop{false};
 };